feat - 세이브로드 작업중 #1

+++ Assets/_LSH/SaveData.cs
@@ -7,5 +7,5 @@ public class SaveData
 {
     //public List<PotInstance> _potInventory = new();
     public PlayerStat _playerStat = new();
-    public float _oxygen;
+    public int _oxygen;
 }
